test: isolate ChatGPT provider migration tests
